fixed compilation error in lte-rrc-header.cc on 32 bit systems

+++ src/lte/model/lte-rrc-header.cc
@@ -1561,7 +1561,7 @@ LteRrcSap::RrcConnectionRequest
 RrcConnectionRequestHeader::GetMessage () const
 {
   RrcConnectionRequest msg;
-  msg.ueIdentity = (m_mmec.to_ulong () << 32) | (m_mTmsi.to_ulong ());
+  msg.ueIdentity = (((uint64_t) m_mmec.to_ulong ()) << 32) | (m_mTmsi.to_ulong ());
 
   return msg;
 }
